pyproject.toml
[project]
name = "rag-trafico-vlc"
version = "0.1.0"
description = "Sistema RAG para consultar el estado del tráfico de Valencia en tiempo real"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "llama-index",
    "llama-index-vector-stores-qdrant",
    "qdrant-client",
    "google-generativeai",
    "tenacity",
    "apscheduler",
    "httpx",
    "python-dotenv",
    "pydantic",
    "llama-index-embeddings-google-genai>=0.5.0",
    "llama-index-llms-google-genai>=0.9.0",
]
